mcp-server-lab0-aoai-response-api/pyproject.toml
[project]
name = "aoai-response-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.1",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
]
